[TS-252] Refactor: select character modal name change
