+++ src/constants/modalConstants.ts
@@ -1,4 +1,4 @@
 export const modalType = {
 	LEAVE_REASON: "LEAVE_REASON",
-	CHARACTER_SELECT: "CHARACTER_SELECT",
+	SELECT_CHARACTER: "SELECT_CHARACTER",
 };
